Refactor logs behind loggers package
